+++ include/acceleration/GridCellArray_t.h
@@ -31,16 +31,16 @@ namespace APTracer { namespace Acceleration {
              * 
              * @param size Initial reserved size of the shape array.
              */
-            GridCellArray_t(unsigned int size);
+            GridCellArray_t(size_t size);
 
             /**
              * @brief Destroy the GridCellArray_t object, destroying its array of shapes.
              */
             virtual ~GridCellArray_t() final;
 
-            unsigned int size_; /**< @brief Size of the array. Number of shapes that can be held by the cell without reallocating.*/
+            size_t size_; /**< @brief Size of the array. Number of shapes that can be held by the cell without reallocating.*/
             Shape_t** items_; /**< @brief Array of shapes contained in the cell. This allows fast iterating and lowest memory use.*/
-            unsigned int increment_size_; /**< @brief Size that will be allocated next time 'reserve()' is called. Increased with 'operator++'.*/
+            size_t increment_size_; /**< @brief Size that will be allocated next time 'reserve()' is called. Increased with 'operator++'.*/
 
             virtual Shape_t* intersect(const Ray_t &ray, double &t, double (&uv)[2]) const final;
             virtual void add(Shape_t* item) final;
