+++ src/acceleration/GridCellArray_t.cpp
@@ -8,7 +8,7 @@ GridCellArray_t::GridCellArray_t() : size_(0), items_(nullptr), increment_size_(
     n_obj_ = 0;
 }
 
-GridCellArray_t::GridCellArray_t(unsigned int size) : size_(size), items_(new Shape_t*[size_]), increment_size_(size_) {
+GridCellArray_t::GridCellArray_t(size_t size) : size_(size), items_(new Shape_t*[size_]), increment_size_(size_) {
     n_obj_ = 0;
 }
 
@@ -25,7 +25,7 @@ Shape_t* GridCellArray_t::intersect(const Ray_t &ray, double &t, double (&uv)[2]
     Shape_t* hit_obj = nullptr; // dunno if this is needed    
     t = std::numeric_limits<double>::infinity();
 
-    for (unsigned int i = 0; i < n_obj_; i++){
+    for (size_t i = 0; i < n_obj_; i++){
         if (items_[i]->intersection(ray, t_temp, uv_temp) && (t_temp < t)){
             hit_obj = items_[i];
             uv[0] = uv_temp[0];
@@ -42,7 +42,7 @@ void GridCellArray_t::add(Shape_t* item){
     }
     else {
         Shape_t** items2 = new Shape_t*[n_obj_+1];
-        for (unsigned int i = 0; i < n_obj_; i++){
+        for (size_t i = 0; i < n_obj_; i++){
             items2[i] = items_[i];
         }
         items2[n_obj_] = item;
@@ -58,9 +58,9 @@ void GridCellArray_t::add(Shape_t* item){
 }
 
 void GridCellArray_t::remove(const Shape_t* item){
-    for (unsigned int i = 0; i < n_obj_; i++){
+    for (size_t i = 0; i < n_obj_; i++){
         if (items_[i] == item){
-            for (unsigned int j = i+1; j < n_obj_; ++j){
+            for (size_t j = i+1; j < n_obj_; ++j){
                 items_[j-1] = items_[j]; 
             }
             --n_obj_;
@@ -76,7 +76,7 @@ void GridCellArray_t::move(Shape_t* item) {
 void GridCellArray_t::reserve(){
     if (n_obj_ > 0){
         Shape_t** items_temp = new Shape_t*[increment_size_];
-        for (unsigned int i = 0; i < n_obj_; ++i){
+        for (size_t i = 0; i < n_obj_; ++i){
             items_temp[i] = items_[i];
         }
         delete [] items_;
